feat(HNG-202): fix folder structure
